Don't recreate cache if nobody listen for reloads.
If nobody listens for reload (no notifiers was added) it will mean the
caller thinks cache is static and the most probably will even not keep
a reference on root. Therefore if we unref it on reload then data that
is received by caller will become invalid.
This sure is a problem of caller but previous design allowed this and
therefore we should add a workarond on this.
The commit adds check for caller which is not aware of any possible
reloads - don't change the cache on server request if there is no notifiers.