Call g_source_remove when destoying channel
ref: https://github.com/lxde/menu-cache/issues/7

It seems that even if the client side of *socket* is already
closed, the server side cannot receive POLLHUP. And when
it already happened, at the time the server tries to write
into the socket, SIGPIPE is received, then we firstly know that
the client is already closed.

Also, then closing GIOChannel, we must explicitly remove the
"source" added by g_io_add_watch, otherwise we see 100 percent
CPU usage by poll.

So we must register source id as well as channel simultaneously,
and call shutdown function explicitly when we receive
SIGPIPE.